+++ openapi/openapi.json
@@ -1049,7 +1049,7 @@
             "type": "integer",
             "example": 100
           },
-          "sumPaymnents": {
+          "sumPayments": {
             "format": "double",
             "type": "number",
             "example": 100.9
