+++ src/main/java/it/gov/pagopa/fdr/repository/fdr/FdrPaymentPublishEntity.java
@@ -1,11 +1,43 @@
 package it.gov.pagopa.fdr.repository.fdr;
 
+import io.quarkus.mongodb.panache.PanacheMongoEntity;
 import io.quarkus.mongodb.panache.common.MongoEntity;
-import it.gov.pagopa.fdr.repository.fdr.model.AbstractReportingFlowPaymentEntity;
+import it.gov.pagopa.fdr.repository.fdr.model.PaymentStatusEnumEntity;
+import java.time.Instant;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
+import org.bson.codecs.pojo.annotations.BsonProperty;
+import org.bson.types.ObjectId;
 
 @Data
 @EqualsAndHashCode(callSuper = true)
 @MongoEntity(collection = "fdr_payment_publish")
-public class FdrPaymentPublishEntity extends AbstractReportingFlowPaymentEntity {}
+public class FdrPaymentPublishEntity extends PanacheMongoEntity {
+
+  private Long revision;
+
+  private Instant created;
+
+  private Instant updated;
+
+  private String iuv;
+  private String iur;
+
+  private Long index;
+  private Double pay;
+
+  @BsonProperty("pay_status")
+  private PaymentStatusEnumEntity payStatus;
+
+  @BsonProperty("pay_date")
+  private Instant payDate;
+
+  @BsonProperty("ref_fdr_id")
+  private ObjectId refFdrId;
+
+  @BsonProperty("ref_fdr_reporting_flow_name")
+  private String refFdrReportingFlowName;
+
+  @BsonProperty("ref_fdr_reporting_sender_psp_id")
+  private String refFdrReportingSenderPspId;
+}
