+++ src/main/java/it/gov/pagopa/fdr/repository/fdr/FdrPublishEntity.java
@@ -1,11 +1,54 @@
 package it.gov.pagopa.fdr.repository.fdr;
 
+import io.quarkus.mongodb.panache.PanacheMongoEntity;
 import io.quarkus.mongodb.panache.common.MongoEntity;
-import it.gov.pagopa.fdr.repository.fdr.model.AbstractReportingFlowEntity;
+import it.gov.pagopa.fdr.repository.fdr.model.ReceiverEntity;
+import it.gov.pagopa.fdr.repository.fdr.model.ReportingFlowStatusEnumEntity;
+import it.gov.pagopa.fdr.repository.fdr.model.SenderEntity;
+import java.time.Instant;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
+import org.bson.codecs.pojo.annotations.BsonProperty;
 
 @Data
 @EqualsAndHashCode(callSuper = true)
 @MongoEntity(collection = "fdr_publish")
-public class FdrPublishEntity extends AbstractReportingFlowEntity {}
+public class FdrPublishEntity extends PanacheMongoEntity {
+
+  private Long revision;
+
+  private Instant created;
+
+  private Instant updated;
+
+  @BsonProperty("reporting_flow_name")
+  private String reportingFlowName;
+
+  @BsonProperty("reporting_flow_date")
+  private Instant reportingFlowDate;
+
+  private SenderEntity sender;
+
+  private ReceiverEntity receiver;
+
+  private String regulation;
+
+  @BsonProperty("regulation_date")
+  private Instant regulationDate;
+
+  @BsonProperty("bic_code_pouring_bank")
+  private String bicCodePouringBank;
+
+  private ReportingFlowStatusEnumEntity status;
+
+  @BsonProperty("tot_payments")
+  private Long totPayments;
+
+  @BsonProperty("sum_payments")
+  private Double sumPayments;
+
+  @BsonProperty("internal_ndp_read")
+  private Boolean internalNdpRead;
+
+  private Boolean read;
+}
