+++ src/main/java/it/gov/pagopa/fdr/service/dto/ReportingFlowGetDto.java
@@ -48,5 +48,5 @@ public class ReportingFlowGetDto {
 
   private Long totPayments;
 
-  private Double sumPaymnents;
+  private Double sumPayments;
 }
