+++ src/main/resources/application.properties
@@ -62,7 +62,7 @@ quarkus.smallrye-openapi.servers=\\$\{host\}/fdr/api/v1 - APIM
 ###################
 quarkus.mongodb.connection-string=${MONGODB_CONNECTION_STRING}
 %dev.quarkus.mongodb.connection-string=mongodb://root:example@localhost:27017/
-%test.quarkus.mongodb.connection-string=mongodb://root:example@localhost:27017/
+#%test.quarkus.mongodb.connection-string=mongodb://root:example@localhost:27017/
 quarkus.mongodb.database=fdr
 
 ###################
