+++ src/test/java/it/gov/pagopa/fdr/rest/organizations/OrganizationResourceTest.java
@@ -1,44 +1,33 @@
 package it.gov.pagopa.fdr.rest.organizations;
 
 import static io.restassured.RestAssured.given;
+import static it.gov.pagopa.fdr.ConstantsTest.pspCode;
+import static it.gov.pagopa.fdr.ConstantsTest.reportingFlowName;
+import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.containsString;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializationFeature;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+import io.quarkiverse.mockserver.test.MockServerTestResource;
+import io.quarkus.test.common.QuarkusTestResource;
 import io.quarkus.test.junit.QuarkusTest;
-import io.quarkus.test.junit.mockito.InjectMock;
 import io.restassured.http.Header;
-import it.gov.pagopa.fdr.Config;
-import it.gov.pagopa.fdr.service.dto.FlowDto;
-import it.gov.pagopa.fdr.service.dto.MetadataDto;
-import it.gov.pagopa.fdr.service.dto.ReceiverDto;
-import it.gov.pagopa.fdr.service.dto.ReportingFlowByIdEcDto;
-import it.gov.pagopa.fdr.service.dto.ReportingFlowGetDto;
-import it.gov.pagopa.fdr.service.dto.ReportingFlowStatusEnumDto;
-import it.gov.pagopa.fdr.service.dto.SenderDto;
-import it.gov.pagopa.fdr.service.dto.SenderTypeEnumDto;
-import it.gov.pagopa.fdr.service.organizations.OrganizationsService;
-import java.time.Instant;
-import java.util.Collections;
-import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.Map;
-import org.junit.jupiter.api.BeforeEach;
+import it.gov.pagopa.fdr.rest.BaseResourceTest;
+import it.gov.pagopa.fdr.rest.organizations.response.GetAllResponse;
+import it.gov.pagopa.fdr.util.MongoResource;
+import it.gov.pagopa.fdr.util.TestUtil;
+import jakarta.inject.Inject;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
-import org.openapi.quarkus.api_config_cache_json.model.ConfigDataV1;
-import org.openapi.quarkus.api_config_cache_json.model.CreditorInstitution;
-import org.openapi.quarkus.api_config_cache_json.model.PaymentServiceProvider;
 
 @QuarkusTest
-public class OrganizationResourceTest {
+@QuarkusTestResource(MockServerTestResource.class)
+@QuarkusTestResource(MongoResource.class)
+public class OrganizationResourceTest extends BaseResourceTest {
 
-  private static final String reportingFlowName = "2016-08-16pspTest-1176";
-  private static final String pspCode = "pspTest";
-  private static final String pspCode2 = "pspTest2";
+  @Inject TestUtil testUtil;
+
   private static final String pspCodeNotEnabled = "pspNotEnabled";
   private static final String brokerCode = "intTest";
   private static final String channelCode = "canaleTest";
@@ -46,104 +35,59 @@ public class OrganizationResourceTest {
   private static final String ecCodeNotEnabled = "00987654321";
   private static final Header header = new Header("Content-Type", "application/json");
 
-  private static final String organizationFindByIdEcUrl =
-      "/organizations/%s/flows?idPsp=%s&page=%d&size=%d";
+  private static final String organizationFindByIdEcUrl = "/organizations/%s/flows?idPsp=%s";
   private static final String organizationfindByReportingFlowNameUrl =
       "/organizations/%s/flows/%s/psps/%s";
 
-  private static ReportingFlowByIdEcDto reportingFlowByIdEc =
-      ReportingFlowByIdEcDto.builder()
-          .count(1L)
-          .data(List.of(FlowDto.builder().name(reportingFlowName).pspId(pspCode).build()))
-          .metadata(MetadataDto.builder().totPage(1).pageSize(10).pageNumber(1).build())
-          .build();
-
-  private static ReportingFlowByIdEcDto reportingFlowByIdEcNoResults =
-      ReportingFlowByIdEcDto.builder()
-          .count(0L)
-          .data(Collections.emptyList())
-          .metadata(MetadataDto.builder().totPage(1).pageSize(10).pageNumber(1).build())
-          .build();
-
-  private static ReportingFlowGetDto reportingFlowGet =
-      ReportingFlowGetDto.builder()
-          .revision(1L)
-          .created(Instant.now())
-          .updated(Instant.now())
-          .status(ReportingFlowStatusEnumDto.CREATED)
-          .reportingFlowName(reportingFlowName)
-          .reportingFlowDate(Instant.parse("2023-04-05T09:21:37.810000Z"))
-          .sender(
-              SenderDto.builder()
-                  .type(SenderTypeEnumDto.LEGAL_PERSON)
-                  .id("SELBIT2B")
-                  .pspId(pspCode)
-                  .pspName("Bank")
-                  .brokerId(brokerCode)
-                  .channelId(channelCode)
-                  .password("1234567890")
-                  .build())
-          .receiver(ReceiverDto.builder().ecId(ecCode).ecName("Comune di xyz").build())
-          .regulation("SEPA - Bonifico xzy")
-          .regulationDate(Instant.parse("2023-04-03T12:00:30.900000Z"))
-          .bicCodePouringBank("UNCRITMMXXX")
-          .build();
-
-  @InjectMock Config config;
-
-  @InjectMock OrganizationsService organizationsService;
-
-  ObjectMapper mapper;
-
-  @BeforeEach
-  void setup() {
-    mapper =
-        new ObjectMapper()
-            .registerModule(new JavaTimeModule())
-            .configure(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS, false);
-    Mockito.doNothing().when(config).init();
-    Mockito.when(config.getClonedCache()).thenReturn(getConfig());
-  }
+  private static String responseAllPublishedFlows = """
+      {
+          "metadata": {
+              "pageSize": 50,
+              "pageNumber": 1,
+              "totPage": 1
+          },
+          "count": 1,
+          "data": [
+              {
+                  "name": "%s",
+                  "pspId": "%s"
+              }
+          ]
+      }""";
 
   /** ############### findByIdEc ################ */
   @Test
   @DisplayName("ORGANIZATIONS findByIdEc Ok")
   public void testOrganization_findByIdEc_Ok() throws JsonProcessingException {
-    Mockito.when(
-            organizationsService.findByIdEc(
-                Mockito.anyString(), Mockito.anyString(), Mockito.anyLong(), Mockito.anyLong()))
-        .thenReturn(reportingFlowByIdEc);
-
-    String url = organizationFindByIdEcUrl.formatted(ecCode, pspCode, 10, 10);
-    ObjectMapper mapper = new ObjectMapper();
-    mapper.registerModule(new JavaTimeModule());
-    given()
+    String flowName = getFlowName();
+    pspSunnyDay(flowName);
+    String url = organizationFindByIdEcUrl.formatted(ecCode, pspCode);
+    String responseFmt = responseAllPublishedFlows.formatted(flowName, pspCode);
+    GetAllResponse res = given()
         .header(header)
         .when()
         .get(url)
         .then()
         .statusCode(200)
-        .body(containsString(mapper.writeValueAsString(reportingFlowByIdEc)));
+        .extract()
+        .as(GetAllResponse.class);
+    assertThat(testUtil.prettyPrint(res), equals(testUtil.prettyPrint(responseFmt, GetAllResponse.class)));
+//        .body(containsString(responseFmt.replaceAll("[\r\n]+"," ").trim()));
   }
 
   @Test
   @DisplayName("ORGANIZATIONS findByIdEc no results")
   public void testOrganization_findByIdEc_OkNoResults() throws JsonProcessingException {
-    Mockito.when(
-            organizationsService.findByIdEc(
-                Mockito.anyString(), Mockito.anyString(), Mockito.anyLong(), Mockito.anyLong()))
-        .thenReturn(reportingFlowByIdEcNoResults);
-
     String url = organizationFindByIdEcUrl.formatted(ecCode, pspCode, 10, 10);
     ObjectMapper mapper = new ObjectMapper();
     mapper.registerModule(new JavaTimeModule());
-    given()
-        .header(header)
-        .when()
-        .get(url)
-        .then()
-        .statusCode(200)
-        .body(containsString(mapper.writeValueAsString(reportingFlowByIdEcNoResults)));
+//    given()
+//        .header(header)
+//        .when()
+//        .get(url)
+//        .then()
+//        .statusCode(200)
+//        .body(containsString(mapper.writeValueAsString(reportingFlowByIdEcNoResults)));
   }
 
   @Test
@@ -196,54 +140,16 @@ public class OrganizationResourceTest {
   @Test
   @DisplayName("ORGANIZATIONS findByReportingFlowName Ok")
   public void testOrganization_findByReportingFlowName_Ok() throws JsonProcessingException {
-    Mockito.when(
-            organizationsService.findByReportingFlowName(Mockito.anyString(), Mockito.anyString()))
-        .thenReturn(reportingFlowGet);
     String url =
         organizationfindByReportingFlowNameUrl.formatted(ecCode, reportingFlowName, pspCode);
 
-    given()
-        .header(header)
-        .when()
-        .get(url)
-        .then()
-        .statusCode(200)
-        .body(containsString(mapper.writeValueAsString(reportingFlowGet)));
+//    given()
+//        .header(header)
+//        .when()
+//        .get(url)
+//        .then()
+//        .statusCode(200)
+//        .body(containsString(mapper.writeValueAsString(reportingFlowGet)));
   }
 
-  private static ConfigDataV1 getConfig() {
-    PaymentServiceProvider paymentServiceProvider = new PaymentServiceProvider();
-    paymentServiceProvider.setEnabled(true);
-    paymentServiceProvider.setPspCode(pspCode);
-
-    PaymentServiceProvider paymentServiceProviderNotEnabled = new PaymentServiceProvider();
-    paymentServiceProviderNotEnabled.setEnabled(false);
-    paymentServiceProviderNotEnabled.setPspCode(pspCodeNotEnabled);
-
-    PaymentServiceProvider paymentServiceProvider2 = new PaymentServiceProvider();
-    paymentServiceProvider2.setEnabled(true);
-    paymentServiceProvider2.setPspCode(pspCode2);
-
-    Map<String, PaymentServiceProvider> psps = new LinkedHashMap<>();
-    psps.put(pspCode, paymentServiceProvider);
-    psps.put(pspCodeNotEnabled, paymentServiceProviderNotEnabled);
-
-    CreditorInstitution creditorInstitution = new CreditorInstitution();
-    creditorInstitution.setCreditorInstitutionCode(ecCode);
-    creditorInstitution.setEnabled(true);
-
-    CreditorInstitution creditorInstitutionNotEnabled = new CreditorInstitution();
-    creditorInstitutionNotEnabled.setCreditorInstitutionCode(ecCodeNotEnabled);
-    creditorInstitutionNotEnabled.setEnabled(false);
-
-    Map<String, CreditorInstitution> creditorInstitutionMap = new LinkedHashMap<>();
-    creditorInstitutionMap.put(ecCode, creditorInstitution);
-    creditorInstitutionMap.put(ecCodeNotEnabled, creditorInstitutionNotEnabled);
-
-    ConfigDataV1 configDataV1 = new ConfigDataV1();
-    configDataV1.setPsps(psps);
-    configDataV1.setCreditorInstitutions(creditorInstitutionMap);
-
-    return configDataV1;
-  }
 }
