+++ src/test/java/it/gov/pagopa/fdr/rest/psps/PspResourceIT.java
@@ -6,6 +6,4 @@ import io.quarkus.test.junit.QuarkusIntegrationTest;
 public class PspResourceIT extends PspResourceTest {
   // Execute the same tests but in packaged mode.
 
-  @Override
-  public void setup() {}
 }
