+++ src/test/java/it/gov/pagopa/fdr/rest/psps/PspResourceTest.java
@@ -1,107 +1,40 @@
 package it.gov.pagopa.fdr.rest.psps;
 
 import static io.restassured.RestAssured.given;
+import static it.gov.pagopa.fdr.ConstantsTest.brokerCode;
+import static it.gov.pagopa.fdr.ConstantsTest.brokerCode2;
+import static it.gov.pagopa.fdr.ConstantsTest.brokerCodeNotEnabled;
+import static it.gov.pagopa.fdr.ConstantsTest.channelCode;
+import static it.gov.pagopa.fdr.ConstantsTest.channelCodeNotEnabled;
+import static it.gov.pagopa.fdr.ConstantsTest.ecCode;
+import static it.gov.pagopa.fdr.ConstantsTest.ecCodeNotEnabled;
+import static it.gov.pagopa.fdr.ConstantsTest.flowsUrl;
+import static it.gov.pagopa.fdr.ConstantsTest.header;
+import static it.gov.pagopa.fdr.ConstantsTest.pspCode;
+import static it.gov.pagopa.fdr.ConstantsTest.pspCode2;
+import static it.gov.pagopa.fdr.ConstantsTest.pspCodeNotEnabled;
+import static it.gov.pagopa.fdr.ConstantsTest.reportingFlowName;
+import static it.gov.pagopa.fdr.ConstantsTest.reportingFlowNameDateWrongFormat;
+import static it.gov.pagopa.fdr.ConstantsTest.reportingFlowNamePspWrongFormat;
 import static org.hamcrest.Matchers.containsString;
 
 import io.quarkiverse.mockserver.test.MockServerTestResource;
 import io.quarkus.test.common.QuarkusTestResource;
 import io.quarkus.test.junit.QuarkusTest;
-import io.restassured.http.Header;
+import it.gov.pagopa.fdr.rest.BaseResourceTest;
 import it.gov.pagopa.fdr.util.MongoResource;
-import java.util.LinkedHashMap;
-import java.util.Map;
-import java.util.Random;
-import java.util.random.RandomGenerator;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.openapi.quarkus.api_config_cache_json.model.BrokerPsp;
-import org.openapi.quarkus.api_config_cache_json.model.Channel;
-import org.openapi.quarkus.api_config_cache_json.model.ConfigDataV1;
-import org.openapi.quarkus.api_config_cache_json.model.CreditorInstitution;
-import org.openapi.quarkus.api_config_cache_json.model.PaymentServiceProvider;
-import org.openapi.quarkus.api_config_cache_json.model.PspChannelPaymentType;
 
 @QuarkusTest
 @QuarkusTestResource(MockServerTestResource.class)
 @QuarkusTestResource(MongoResource.class)
-public class PspResourceTest {
-
-  private static final String reportingFlowName = "2016-08-16pspTest-1176";
-  private static final String reportingFlowNamePspWrongFormat = "2016-08-16-psp-1176";
-  private static final String reportingFlowNameDateWrongFormat = "2016-aa-16pspTest-1176";
-  private static final String pspCode = "pspTest";
-  private static final String pspCode2 = "pspTest2";
-  private static final String pspCodeNotEnabled = "pspNotEnabled";
-  private static final String brokerCode = "intTest";
-  private static final String brokerCode2 = "intTest2";
-  private static final String brokerCodeNotEnabled = "intNotEnabled";
-  private static final String channelCode = "canaleTest";
-  private static final String channelCodeNotEnabled = "canaleNotEnabled";
-  private static final String ecCode = "12345678900";
-  private static final String ecCodeNotEnabled = "00987654321";
-  private static final String pspChannelPaymentTypeCode = "PAYPAL";
-  private static final Header header = new Header("Content-Type", "application/json");
-
-  private static final String flowsUrl = "/psps/%s/flows";
-
-  private static String template = """
-        {
-          "reportingFlowName": "%s",
-          "reportingFlowDate": "2023-04-05T09:21:37.810000Z",
-          "sender": {
-            "type": "LEGAL_PERSON",
-            "id": "SELBIT2B",
-            "pspId": "%s",
-            "pspName": "Bank",
-            "brokerId": "%s",
-            "channelId": "%s",
-            "password": "1234567890"
-          },
-          "receiver": {
-            "id": "APPBIT2B",
-            "ecId": "%s",
-            "ecName": "Comune di xyz"
-          },
-          "regulation": "SEPA - Bonifico xzy",
-          "regulationDate": "2023-04-03T12:00:30.900000Z",
-          "bicCodePouringBank": "UNCRITMMXXX"
-        }
-        """;
-
-  String response = """
-      {"message":"Flow [%s] saved"}""";
-
-//  @InjectMock
-//  PspsService pspsService;
-
-
-  @BeforeEach
-  public void setup() {
-//    Mockito.doNothing().when(config).init();
-//    Mockito.when(config.getClonedCache()).thenReturn(getConfig());
-
-//    Mockito.doNothing().when(pspsService).save(null);
-  }
+public class PspResourceTest extends BaseResourceTest {
 
   @Test
-  @DisplayName("PSPS create OK")
+  @DisplayName("PSPS publish OK")
   public void testPspOk() {
-    RandomGenerator randomGenerator = new Random();
-    String flowName = reportingFlowName.substring(0, reportingFlowName.length()-4)+randomGenerator.nextInt(1111,9999);
-    String url = flowsUrl.formatted(pspCode);
-    String bodyFmt =
-        template.formatted(flowName, pspCode, brokerCode, channelCode, ecCode);
-    String responseFmt = response.formatted(flowName);
-
-    given()
-        .body(bodyFmt)
-        .header(header)
-        .when()
-        .post(url)
-        .then()
-        .statusCode(201)
-        .body(containsString(responseFmt));
+    pspSunnyDay(getFlowName());
   }
 
   @Test
@@ -111,7 +44,7 @@ public class PspResourceTest {
 
     String url = flowsUrl.formatted(pspCode);
     String bodyFmt =
-        template.formatted(reportingFlowName, pspNotMatch, brokerCode, channelCode, ecCode);
+        flowTemplate.formatted(reportingFlowName, pspNotMatch, brokerCode, channelCode, ecCode);
     String responseFmt =
         """
         {"httpStatusCode":400,"httpStatusDescription":"Bad Request","appErrorCode":"FDR-0704","errors":[{"message":"Reporting flow [2016-08-16pspTest-1176] have sender.pspId [PSP_NOT_MATCH] but not match with query param [pspTest]"}]}""";
@@ -134,7 +67,7 @@ public class PspResourceTest {
 
     String url = flowsUrl.formatted(pspUnknown);
     String bodyFmt =
-        template.formatted(reportingFlowName, pspUnknown, brokerCode, channelCode, ecCode);
+        flowTemplate.formatted(reportingFlowName, pspUnknown, brokerCode, channelCode, ecCode);
     String responseFmt =
         """
         {"httpStatusCode":400,"httpStatusDescription":"Bad Request","appErrorCode":"FDR-0708","errors":[{"message":"Psp [PSP_UNKNOWN] unknown"}]}""";
@@ -156,7 +89,7 @@ public class PspResourceTest {
 
     String url = "/psps/%s/flows".formatted(pspCodeNotEnabled);
     String bodyFmt =
-        template.formatted(reportingFlowName, pspCodeNotEnabled, brokerCode, channelCode, ecCode);
+        flowTemplate.formatted(reportingFlowName, pspCodeNotEnabled, brokerCode, channelCode, ecCode);
     String responseFmt =
         """
         {"httpStatusCode":400,"httpStatusDescription":"Bad Request","appErrorCode":"FDR-0709","errors":[{"message":"Psp [pspNotEnabled] not enabled"}]}""";
@@ -178,7 +111,7 @@ public class PspResourceTest {
 
     String url = flowsUrl.formatted(pspCode);
     String bodyFmt =
-        template.formatted(reportingFlowName, pspCode, brokerPspUnknown, channelCode, ecCode);
+        flowTemplate.formatted(reportingFlowName, pspCode, brokerPspUnknown, channelCode, ecCode);
     String responseFmt =
         """
         {"httpStatusCode":400,"httpStatusDescription":"Bad Request","appErrorCode":"FDR-0710","errors":[{"message":"Broker [BROKERPSP_UNKNOWN] unknown"}]}""";
@@ -198,7 +131,7 @@ public class PspResourceTest {
   public void test_brokerpsp_KO_FDR0711() {
     String url = flowsUrl.formatted(pspCode);
     String bodyFmt =
-        template.formatted(reportingFlowName, pspCode, brokerCodeNotEnabled, channelCode, ecCode);
+        flowTemplate.formatted(reportingFlowName, pspCode, brokerCodeNotEnabled, channelCode, ecCode);
     String responseFmt =
         """
         {"httpStatusCode":400,"httpStatusDescription":"Bad Request","appErrorCode":"FDR-0711","errors":[{"message":"Broker [intNotEnabled] not enabled"}]}""";
@@ -220,7 +153,7 @@ public class PspResourceTest {
 
     String url = flowsUrl.formatted(pspCode);
     String bodyFmt =
-        template.formatted(reportingFlowName, pspCode, brokerCode, channelUnknown, ecCode);
+        flowTemplate.formatted(reportingFlowName, pspCode, brokerCode, channelUnknown, ecCode);
     String responseFmt =
         """
         {"httpStatusCode":400,"httpStatusDescription":"Bad Request","appErrorCode":"FDR-0712","errors":[{"message":"Channel [CHANNEL_UNKNOWN] unknown"}]}""";
@@ -240,7 +173,7 @@ public class PspResourceTest {
   public void test_channel_KO_FDR0713() {
     String url = flowsUrl.formatted(pspCode);
     String bodyFmt =
-        template.formatted(reportingFlowName, pspCode, brokerCode, channelCodeNotEnabled, ecCode);
+        flowTemplate.formatted(reportingFlowName, pspCode, brokerCode, channelCodeNotEnabled, ecCode);
     String responseFmt =
         """
         {"httpStatusCode":400,"httpStatusDescription":"Bad Request","appErrorCode":"FDR-0713","errors":[{"message":"channelId.notEnabled"}]}""";
@@ -260,7 +193,7 @@ public class PspResourceTest {
   public void test_channelBroker_KO_FDR0714() {
     String url = flowsUrl.formatted(pspCode);
     String bodyFmt =
-        template.formatted(reportingFlowName, pspCode, brokerCode2, channelCode, ecCode);
+        flowTemplate.formatted(reportingFlowName, pspCode, brokerCode2, channelCode, ecCode);
     String responseFmt =
         """
         {"httpStatusCode":400,"httpStatusDescription":"Bad Request","appErrorCode":"FDR-0714","errors":[{"message":"Channel [canaleTest] with broker [intTest2] not authorized"}]}""";
@@ -280,7 +213,7 @@ public class PspResourceTest {
   public void test_channelPsp_KO_FDR0715() {
     String url = flowsUrl.formatted(pspCode2);
     String bodyFmt =
-        template.formatted(reportingFlowName, pspCode2, brokerCode, channelCode, ecCode);
+        flowTemplate.formatted(reportingFlowName, pspCode2, brokerCode, channelCode, ecCode);
     String responseFmt =
         """
         {"httpStatusCode":400,"httpStatusDescription":"Bad Request","appErrorCode":"FDR-0715","errors":[{"message":"Channel [canaleTest] with psp [pspTest2] not authorized"}]}""";
@@ -301,7 +234,7 @@ public class PspResourceTest {
     String ecUnknown = "EC_UNKNOWN";
     String url = flowsUrl.formatted(pspCode);
     String bodyFmt =
-        template.formatted(reportingFlowName, pspCode, brokerCode, channelCode, ecUnknown);
+        flowTemplate.formatted(reportingFlowName, pspCode, brokerCode, channelCode, ecUnknown);
     String responseFmt =
         """
         {"httpStatusCode":400,"httpStatusDescription":"Bad Request","appErrorCode":"FDR-0716","errors":[{"message":"Creditor institution [EC_UNKNOWN] unknown"}]}""";
@@ -321,7 +254,7 @@ public class PspResourceTest {
   public void test_ecId_KO_FDR0717() {
     String url = flowsUrl.formatted(pspCode);
     String bodyFmt =
-        template.formatted(reportingFlowName, pspCode, brokerCode, channelCode, ecCodeNotEnabled);
+        flowTemplate.formatted(reportingFlowName, pspCode, brokerCode, channelCode, ecCodeNotEnabled);
     String responseFmt =
         """
         {"httpStatusCode":400,"httpStatusDescription":"Bad Request","appErrorCode":"FDR-0717","errors":[{"message":"Creditor institution [00987654321] not enabled"}]}""";
@@ -341,7 +274,7 @@ public class PspResourceTest {
   public void test_flowName_KO_FDR0718() {
     String url = flowsUrl.formatted(pspCode);
     String bodyFmt =
-        template.formatted(reportingFlowNameDateWrongFormat, pspCode, brokerCode, channelCode, ecCode);
+        flowTemplate.formatted(reportingFlowNameDateWrongFormat, pspCode, brokerCode, channelCode, ecCode);
     String responseFmt =
         """
         {"httpStatusCode":400,"httpStatusDescription":"Bad Request","appErrorCode":"FDR-0718","errors":[{"message":"Reporting flow [2016-aa-16pspTest-1176] has wrong date"}]}""";
@@ -361,7 +294,7 @@ public class PspResourceTest {
   public void test_flowName_KO_FDR0719() {
     String url = flowsUrl.formatted(pspCode);
     String bodyFmt =
-        template.formatted(reportingFlowNamePspWrongFormat, pspCode, brokerCode, channelCode, ecCode);
+        flowTemplate.formatted(reportingFlowNamePspWrongFormat, pspCode, brokerCode, channelCode, ecCode);
     String responseFmt =
         """
         {"httpStatusCode":400,"httpStatusDescription":"Bad Request","appErrorCode":"FDR-0719","errors":[{"message":"Reporting flow [2016-08-16-psp-1176] has wrong psp"}]}""";
@@ -376,81 +309,4 @@ public class PspResourceTest {
         .body(containsString(responseFmt));
   }
 
-
-  private static ConfigDataV1 getConfig() {
-    PaymentServiceProvider paymentServiceProvider = new PaymentServiceProvider();
-    paymentServiceProvider.setEnabled(true);
-    paymentServiceProvider.setPspCode(pspCode);
-
-    PaymentServiceProvider paymentServiceProviderNotEnabled = new PaymentServiceProvider();
-    paymentServiceProviderNotEnabled.setEnabled(false);
-    paymentServiceProviderNotEnabled.setPspCode(pspCodeNotEnabled);
-
-    PaymentServiceProvider paymentServiceProvider2 = new PaymentServiceProvider();
-    paymentServiceProvider2.setEnabled(true);
-    paymentServiceProvider2.setPspCode(pspCode2);
-
-    Map<String, PaymentServiceProvider> psps = new LinkedHashMap<>();
-    psps.put(pspCode, paymentServiceProvider);
-    psps.put(pspCode2, paymentServiceProvider2);
-    psps.put(pspCodeNotEnabled, paymentServiceProviderNotEnabled);
-
-    BrokerPsp brokerPsp = new BrokerPsp();
-    brokerPsp.setEnabled(true);
-    brokerPsp.setBrokerPspCode(brokerCode);
-
-    BrokerPsp brokerPsp2 = new BrokerPsp();
-    brokerPsp2.setEnabled(true);
-    brokerPsp2.setBrokerPspCode(brokerCode2);
-
-    BrokerPsp brokerPspNotEnabled = new BrokerPsp();
-    brokerPspNotEnabled.setEnabled(false);
-    brokerPspNotEnabled.setBrokerPspCode(brokerCodeNotEnabled);
-
-    Map<String, BrokerPsp> pspBrokers = new LinkedHashMap<>();
-    pspBrokers.put(brokerCode, brokerPsp);
-    pspBrokers.put(brokerCode2, brokerPsp2);
-    pspBrokers.put(brokerCodeNotEnabled, brokerPspNotEnabled);
-
-    Channel channel = new Channel();
-    channel.setEnabled(true);
-    channel.setBrokerPspCode(brokerCode);
-    channel.setChannelCode(channelCode);
-
-    Channel channelNotEnabled = new Channel();
-    channelNotEnabled.setEnabled(false);
-    channelNotEnabled.setBrokerPspCode(brokerCode);
-    channelNotEnabled.setChannelCode(channelCodeNotEnabled);
-
-    Map<String, Channel> channels = new LinkedHashMap<>();
-    channels.put(channelCode, channel);
-    channels.put(channelCodeNotEnabled, channelNotEnabled);
-
-    PspChannelPaymentType pspChannelPaymentType = new PspChannelPaymentType();
-    pspChannelPaymentType.setPspCode(pspCode);
-    pspChannelPaymentType.setChannelCode(channelCode);
-    Map<String, PspChannelPaymentType> pspChannelPaymentTypeLinkedHashMap = new LinkedHashMap<>();
-    pspChannelPaymentTypeLinkedHashMap.put(pspChannelPaymentTypeCode, pspChannelPaymentType);
-
-    CreditorInstitution creditorInstitution = new CreditorInstitution();
-    creditorInstitution.setCreditorInstitutionCode(ecCode);
-    creditorInstitution.setEnabled(true);
-
-    CreditorInstitution creditorInstitutionNotEnabled = new CreditorInstitution();
-    creditorInstitutionNotEnabled.setCreditorInstitutionCode(ecCodeNotEnabled);
-    creditorInstitutionNotEnabled.setEnabled(false);
-
-    Map<String, CreditorInstitution> creditorInstitutionMap = new LinkedHashMap<>();
-    creditorInstitutionMap.put(ecCode, creditorInstitution);
-    creditorInstitutionMap.put(ecCodeNotEnabled, creditorInstitutionNotEnabled);
-
-    ConfigDataV1 configDataV1 = new ConfigDataV1();
-    configDataV1.setPsps(psps);
-    configDataV1.setPspBrokers(pspBrokers);
-    configDataV1.setChannels(channels);
-    configDataV1.setPspChannelPaymentTypes(pspChannelPaymentTypeLinkedHashMap);
-    configDataV1.setCreditorInstitutions(creditorInstitutionMap);
-
-    return configDataV1;
-  }
 }
